Fixed spindle settings and test screens

+++ platform/nucleo-u5a5zj/TouchGFX/gui/include/gui/containers/ValueSelector.hpp
@@ -3,14 +3,14 @@
 
 #include <gui_generated/containers/ValueSelectorBase.hpp>
 
-class ValueSelector : public ValueSelectorBase
-{
+class ValueSelector : public ValueSelectorBase {
 public:
     ValueSelector();
     virtual ~ValueSelector() {}
 
     virtual void initialize();
-    void setDisplayedValue(const uint16_t value);
+    void setDisplayedValue(const uint32_t value);
+
 protected:
 };
